requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
transformers==4.11.3
sentence-transformers==2.1.0
faiss-cpu==1.7.1
numpy==1.21.2
scikit-learn==0.24.2
torch==1.9.0
fastapi-cache2==0.1.8